chore: better TypeGuard for Bundles and BundleCaps as used
